### 2019-08-23 #1678 @whedon
[PRE REVIEW]: pypulseq 
**Submitting author:** @imr-framework (<a href="http://orcid.org/0000-0002-3776-4114">Sairam Geethanath</a>)
**Repository:** <a href="https://github.com/imr-framework/pypulseq" target ="_blank">https://github.com/imr-framework/pypulseq</a>
**Version:** 1.2.2
