### 2019-08-23 #1678 @kyleniemeyer
Hi @Kevin-Mattheus-Moerman, another imaging-related submission—do you have the bandwidth to edit this one?
