### 2019-08-29 #1678 @grlee77
Hi @sravan953, I am interested in taking a look. I work in a Philips environment currently, but have also done sequence programming for GE and Siemens scanners in the past, so this is pretty relevant.
