### 2019-08-29 #1678 @Kevin-Mattheus-Moerman
@mwegrzyn are you interested in reviewing this work for JOSS?
