### 2019-08-30 #1678 @mwegrzyn
Hello,
thank you for the invitation. This sounds very interesting, but I'm afraid I don't have the necessary physics knowledge to adequately understand the code. Sorry, but I'll have to pass.
Best regards, Martin Wegrzyn
