### 2019-08-30 #1678 @Kevin-Mattheus-Moerman
@kelvinlayton @stefan-k would you be interested in reviewing this work for the Journal of Open Source Software (JOSS)? [The review](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html) focuses largely on [the software](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html) and a [very short paper](https://github.com/openjournals/joss-papers/blob/joss.01678/joss.01678/10.21105.joss.01678.pdf).
