### 2019-08-30 #1678 @stefan-k
Thanks for the invitation!
I'm currently quite short on time, but I'll think about it and I'll let you know within the next couple of days. I should mention though that there may be a conflict of interest as I've collaborated with @imr-framework in the past and we've co-authored papers and conference contributions. I've also worked on a [similar project](https://github.com/pulseq/pypulseq) which was abandoned in favor of the software which is up for review.

Maybe @m-a-x-i-m-z is interested in this too.
