### 2019-09-03 #1678 @Kevin-Mattheus-Moerman
@stefan-k thanks for your message and for clarifying the potential conflict of interest. Since we will have multiple reviewers without a conflict of interest, if you were to join decisions will not rest solely on your judgement/comments. Therefore, since it sounds like you could have valuable feedback, I'd still welcome your contribution as an additional reviewer. Let me know what you decide.
