### 2019-09-03 #1678 @stefan-k
Unfortunately I have to decline as I won't have time. Maybe @jfnielsen or @magland, who have worked on similar software, are interested to help out.
