### 2019-09-03 #1678 @jfnielsen
I'm happy to review but like stefan-k I'm collaborating with the authors on related topics which is technically a conflict. I'm certainly interested though.
