### 2019-09-04 #1678 @Kevin-Mattheus-Moerman
@jfnielsen After consulting with the editorial board it was decided to not include you due to the conflict of interest you referred to. My apologies for the confusion. Thanks again for offering to review for JOSS. I will consider you as a potential reviewer for other MRI related submissions.
