### 2019-09-05 #1678 @Kevin-Mattheus-Moerman
@mfroeling are you or your co-workers interested in reviewing this work on MRI pulse sequence design in Python?
