### 2019-09-05 imr-framework/pypulseq#16 @grlee77
adding a test suite and continuous integration
In considering the JOSS review criteria, I think the lack of a test suite or any type of continuous integration to verify correct operation across platforms / software versions is the biggest weakness.

There are fortunately a number of demos that the user can run to verify the basic operation of the package, but I did see any automated way to run tests.

**Describe the solution you'd like**

Ideally some automated unit tests would be created. The [pytest](https://docs.pytest.org/en/latest/) package is a popular choice for this.

Once unit tests have been defined, services like Travis-CI, Circle-CI, Appveyor and/or Azure can be used to set up automated cross-platform testing on each pull request, etc. This may be overkill at the moment for a project of this size, but something to keep in mind.

**Additional context**
