+++ pypulseq/__init__.py
@@ -1,1 +1,2 @@
 name = 'pypulseq'
+__version__ = '1.2.3'
