+++ pypulseq/__init__.py
@@ -1,2 +1,28 @@
 name = 'pypulseq'
 __version__ = '1.2.3'
+__all__ = ['Sequence',
+           'add_gradients',
+           'add_ramps',
+           'align',
+           'calc_duration',
+           'calc_ramp',
+           'calc_rf_center',
+           'check_timing',
+           'compress_shape',
+           'convert',
+           'decompress_shape',
+           'event_lib',
+           'make_adc',
+           'make_arbitrary_grad',
+           'make_arbitrary_rf',
+           'make_block_pulse',
+           'make_delay',
+           'make_extended_trapezoid',
+           'make_gauss_pulse',
+           'make_sinc_pulse',
+           'make_trap_pulse',
+           'opts',
+           'points_to_waveform',
+           'split_gradient',
+           'split_gradient_at',
+           'traj_to_grad']
