### 2019-09-08 373b7f5
Fix #13
All plots share x-axis
