### 2019-09-08 2268bf6
Fix minor bugs
- Fix quantisation factor to 1e-8 from 1e-7 in compress_shape.py
- Print test reports in GRE and TSE
- Fix amplitude in make_trap_pulse.py
