+++ pypulseq/make_trap_pulse.py
@@ -55,11 +55,13 @@ def make_trapezoid(channel='z', system=Opts(), duration=0, area=-1, flat_time=No
         raise ValueError('Must supply either \'area\', \'flat_area\' or \'amplitude\'')
 
     if flat_time is not None:
-        if amplitude is  None:
-            amplitude = flat_area / flat_time
+        if amplitude is not None:
+            amplitude2 = amplitude
+        else:
+            amplitude2 = flat_area / flat_time
 
         if rise_time is None:
-            rise_time = abs(amplitude) / max_slew
+            rise_time = abs(amplitude2) / max_slew
             rise_time = math.ceil(rise_time / system.grad_raster_time) * system.grad_raster_time
         fall_time, flat_time = rise_time, flat_time
     elif duration > 0:
@@ -67,53 +69,53 @@ def make_trapezoid(channel='z', system=Opts(), duration=0, area=-1, flat_time=No
             if rise_time is None:
                 dC = 1 / abs(2 * max_slew) + 1 / abs(2 * max_slew)
                 possible = duration ** 2 > 4 * abs(area) * dC
-                amplitude = (duration - math.sqrt(duration ** 2 - 4 * abs(area) * dC)) / (2 * dC)
+                amplitude2 = (duration - math.sqrt(duration ** 2 - 4 * abs(area) * dC)) / (2 * dC)
             else:
-                amplitude = area / (duration - rise_time)
-                possible = duration > 2 * rise_time and abs(amplitude) < max_grad
+                amplitude2 = area / (duration - rise_time)
+                possible = duration > 2 * rise_time and abs(amplitude2) < max_grad
 
             if not possible:
                 raise ValueError('Requested area is too large for this gradient')
 
         if rise_time is None:
             rise_time = math.ceil(
-                abs(amplitude) / max_slew / system.grad_raster_time) * system.grad_raster_time
+                abs(amplitude2) / max_slew / system.grad_raster_time) * system.grad_raster_time
 
         fall_time = rise_time
         flat_time = duration - rise_time - fall_time
 
         if amplitude is None:
-            amplitude = area / (rise_time / 2 + fall_time / 2 + flat_time)
+            amplitude2 = area / (rise_time / 2 + fall_time / 2 + flat_time)
 
     else:
         if area is None:
             raise ValueError('Must supply a duration')
         else:
             rise_time = math.ceil(math.sqrt(abs(area) / max_slew) / system.grad_raster_time) * system.grad_raster_time
-            amplitude = area / rise_time
+            amplitude2 = area / rise_time
             t_eff = rise_time
 
-            if abs(amplitude) > max_grad:
+            if abs(amplitude2) > max_grad:
                 t_eff = math.ceil(abs(area) / max_grad / system.grad_raster_time) * system.grad_raster_time
-                amplitude = area / t_eff
+                amplitude2 = area / t_eff
                 rise_time = math.ceil(
-                    abs(amplitude) / max_slew / system.grad_raster_time) * system.grad_raster_time
+                    abs(amplitude2) / max_slew / system.grad_raster_time) * system.grad_raster_time
 
             flat_time = t_eff - rise_time
             fall_time = rise_time
 
-    if abs(amplitude) > max_grad:
+    if abs(amplitude2) > max_grad:
         raise ValueError("Amplitude violation")
 
     grad = SimpleNamespace()
     grad.type = 'trap'
     grad.channel = channel
-    grad.amplitude = amplitude
+    grad.amplitude = amplitude2
     grad.rise_time = rise_time
     grad.flat_time = flat_time
     grad.fall_time = fall_time
[…]
