### 2019-09-11 #1678 @Kevin-Mattheus-Moerman
@nstikov, @vsoch, @mathieuboudreau would you be interested in reviewing this JOSS submission?
