### 2019-09-11 #1725 @whedon
[REVIEW]: pypulseq: A Python Package for MRI Pulse Sequence Design 
**Submitting author:** @imr-framework (<a href="http://orcid.org/0000-0002-3776-4114">Sairam Geethanath</a>)
**Repository:** <a href="https://github.com/imr-framework/pypulseq" target ="_blank">https://github.com/imr-framework/pypulseq</a>
**Version:** 1.2.2r1

@grlee77, @mathieuboudreau, and @spinicist, please carry out your review in this issue by updating the checklist below. If you cannot edit the checklist please:
