### 2019-09-13 #1725 @grlee77
I have performed my initial review.

**Summary**

The Pypulseq software is a port of the Pulseq Matlab package to Python. This makes the Pulseq format for rapid sequence prototyping more readily available to sites without need for a commercial license. The authors also have used Python-based machine learning in combination with Pypulseq for novel sequence design. The overall goal is to have a tool for rapid sequence development without relying on complicated and proprietary vendor-specific pulse sequence development tools. The advantage of such an approach is greatly reduced development time for simple experiments and rapid prototyping.

I verified that Pulseq files are produced by running the examples. The generated files are plain text and appear to conform to the Pulseq specification. I currently work with Philips MR systems, which is not one of the supported vendors, so I did not attempt to verify proper execution of the Pulseq files on a physical MRI scanner.

The software appears to work as advertised. Its main weakness in relation to the JOSS criteria is the lack of automated testing and/or continuous integration to test across platforms. I opened imr-framework/pypulseq#16 regarding this.

**Comments regarding the text of the paper**

1.) A brief statement on the limitations of the types of sequences that can be designed with this simplified approach is warranted so as not to oversell the capabilities. For example, it seems that scans requiring features such as cardiac or respiratory gating, feedback from real-time image-based navigators, etc. are not going to be possible. I assume that it also presents difficulty for scans that require more complicated planning such as placement of saturation slabs and or localized shim boxes interactively prior to scan start.

2.) Similarly, is offline image reconstruction (i.e. not with the vendor-supplied reconstructor) always required when using this tool?

3.) I would add a reference to J. Nielsen's TOPPE publication as a closely related work (https://doi.org/10.1002/mrm.26990). (It seems that on the GE platform a Pulseq->TOPPE conversion is done to actually play out the sequences on the scanner?).

4.) For the puposes of the "state of the field" JOSS criteria, it should probably also be mentioned how this framework (and TOPPE) are fast and do not require compilation (as compared to C++ frameworks like ODIN or SequenceTree).

5.) Are there any areas that pypulseq currently offers additional functionality (or lacks functionality) relative to the Matlab implementation? This would be good to know for existing Matlab users thinking of making the switch to Python.

**Minor typos/suggested phrasing changes**

    1.) in the summary:
    "easy integrability with" -> "easy integration with"

    2.) under the "Statement of Need"
    "These are aimed at achieving" <- unclear what "these" is referring to

    3.) in the section on the Pulseq format:

    "A pulse sequence comprises of pulsed, magnetic field gradient, delay or ADC readout events" ->

    "A pulse sequence comprises of radiofrequency pulses, magnetic field gradient waveforms, delays or analog-to-digital converter (ADC) readout events"

    4.) Please include the meeting names (International Society for Magnetic Resonance in Medicine?) for the fourth reference and last reference.
