### 2019-09-16 imr-framework/pypulseq#16 @sravan953
@grlee77 Thank you for your comments regarding automated testing/CI. We have been discussing this internally and plan to incorporate the same in our next major release (2.0.0) as a part of ISMRM 2021 efforts.
