### 2019-09-18 bf9268c
Update README.md, paper.md and paper.bib
Address @grlee77's comments - https://github.com/openjournals/joss-reviews/issues/1725#issuecomment-531279644
