+++ paper.md
@@ -28,61 +28,70 @@ Magnetic Resonance Imaging (MRI) is a critical component of healthcare. MRI data
 radio-frequency and magnetic field gradient pulses. Designing these pulse sequences requires knowledge of specific 
 programming environments depending on the vendor hardware (generations) and software (revisions) intended for 
 implementation. This impedes the pace of prototyping. Pulseq [@layton2017pulseq] introduced an open-source file 
-standard for Siemens/[GE](toppe)/[Bruker](bruker) platforms. In this work, we introduce `pypulseq`, which enables pulse sequence 
-programming in Python. Its advantages are zero licensing fees and easy integrability with deep learning methods 
-developed in Python. `pypulseq` is aimed at MR researchers, faculty, students, and other allied field researchers 
-such as those in neuroscience. We have leveraged this tool for several published research works.
+standard for pulse sequences that can be deployed on Siemens/GE via [TOPPE](toppe) [@nielsen2018toppe]/[Bruker](bruker) 
+platforms. In this work, we introduce `PyPulseq`, which enables pulse sequence programming in Python. Its advantages 
+are zero licensing fees and easy integration with deep learning methods developed in Python. `PyPulseq` is aimed at MR 
+researchers, faculty, students, and other allied field researchers such as those in neuroscience. We have leveraged 
+this tool for several published research works.
 
 # Statement of need
 
 Magnetic Resonance Imaging (MRI) is a non-invasive diagnostic imaging tool. It is a critical component of healthcare 
 and has a significant impact on diagnosis and treatment assessment. Structural, functional and metabolic MRI generate 
-valuable information that aid in the accurate diagnosis of a wide range of pathologies. These are aimed at achieving 
-faster scan times, improving tissue contrast and increasing Signal-to-Noise Ratio (SNR). However, designing pulse 
-sequences requires knowledge of specific programming environments depending on the vendor hardware (generations) and 
-software (revisions) intended for implementation. This typically involves considerable effort, impeding the pace of 
-prototyping and therefore research and development. 
+valuable information that aid in the accurate diagnosis of a wide range of pathologies. A unique strength of MRI is the 
+ability to visualise diverse pathologies achieved by the flexibility in designing tailored pulse sequences. MR pulse 
+sequences are a collection of radio-frequency and gradient waveforms that are executed on the scanner hardware to 
+acquire raw data. 
 
-Typically, MRl researchers program and simulate the pulse sequences on computers and execute them on MR scanners. 
-However, the programming language differs across vendors. This also hampers multi-site multi-vendor studies as it 
-requires researchers to be acquainted with each vendor's programming environment. Further, harmonizing acquisition 
-across vendors in MR is challenging given the sophisticated choices with respect to a particular hardware platform, 
-software and the application. This work introduces an open-source tool that enables pulse sequence programming for 
-Siemens/GE/Bruker platforms in Python.
+Research efforts on pulse sequence design are directed at achieving faster scan times, improving tissue contrast and 
+increasing Signal-to-Noise Ratio (SNR). However, designing pulse sequences requires knowledge of specific programming 
+environments depending on the vendor hardware (generations) and software (revisions) intended for implementation. 
+Typically, MRI researchers program and simulate the pulse sequences on computers and execute them on MR scanners. This 
+typically involves considerable effort, impeding the pace of prototyping and therefore research and development. This 
[…]
