+++ paper.bib
@@ -38,9 +38,42 @@
 @inproceedings{gehua2019ismrm,
   address = {Montreal, Canada},
   author = {Tong, Gehua and Geethanath, Sairam and Qian, Enlin and Ravi, Keerthi Sravan and Jimeno Manso, Marina and Vaughan, John Thomas},
-  booktitle = {ISMRM 27th Annual Meeting & Exhibition},
+  booktitle = {ISMRM 27th Annual Meeting and Exhibition},
   title = {Virtual MR Scanner Software},
   year = {2019}
+}
+
+@article{jochimsen2004odin,
+  title={ODIN—object-oriented development interface for NMR},
+  author={Jochimsen, Thies H and Von Mengershausen, Michael},
+  journal={Journal of Magnetic Resonance},
+  volume={170},
+  number={1},
+  pages={67--78},
+  year={2004},
+  publisher={Elsevier}
+}
+
+@article{magland2016pulse,
+  title={Pulse sequence programming in a dynamic visual environment: SequenceTree},
+  author={Magland, Jeremy F and Li, Cheng and Langham, Michael C and Wehrli, Felix W},
+  journal={Magnetic resonance in medicine},
+  volume={75},
+  number={1},
+  pages={257--265},
+  year={2016},
+  publisher={Wiley Online Library}
+}
+
+@article{nielsen2018toppe,
+  title={TOPPE: A framework for rapid prototyping of MR pulse sequences},
+  author={Nielsen, Jon-Fredrik and Noll, Douglas C},
+  journal={Magnetic resonance in medicine},
+  volume={79},
+  number={6},
+  pages={3128--3134},
+  year={2018},
+  publisher={Wiley Online Library}
 }
 
 @inproceedings{ravi2019accessible-amri,
@@ -54,7 +87,7 @@
 @inproceedings{ravi2019selfadmin,
   address = {Montreal, Canada},
   author = {Ravi, Keerthi Sravan and Geethanath, Sairam and Vaughan, John Thomas},
-  booktitle = {ISMRM 27th Annual Meeting & Exhibition},
+  booktitle = {ISMRM 27th Annual Meeting and Exhibition},
   title = {Self-administered exam using Autonomous Magnetic Resonance Imaging (AMRI)},
   year = {2019}
 }
