+++ README.md
@@ -7,13 +7,13 @@
 Pulse sequence design is a significant component of MRI research. However, multi-vendor studies require researchers to
 be acquainted with each hardware platform's programming environment.
 
-`pypulseq` enables vendor-neutral pulse sequence design in Python [[1]](#references). The pulse sequences can be
+`PyPulseq` enables vendor-neutral pulse sequence design in Python [[1]](#references). The pulse sequences can be
 exported as a `.seq` file to be run on  Siemens/[GE]/[Bruker] hardware by leveraging their respective
 Pulseq interpreters. This tool is targeted at MRI pulse sequence designers, researchers, students and other interested
 users. It is a translation of the Pulseq framework originally written in Matlab [[2]](#references).
 
 It is strongly recommended to first read the [Pulseq specification]  before proceeding. The specification
-document defines the concepts required for pulse sequence design using `pypulseq`. API docs can be found [here][api-docs].
+document defines the concepts required for pulse sequence design using `PyPulseq`. API docs can be found [here][api-docs].
 
 ## Dependencies
 1. numpy>=1.16.3
@@ -28,7 +28,7 @@ document defines the concepts required for pulse sequence design using `pypulseq
 ?. Get in touch regarding running the `.seq` files on your Siemens/[GE]/[Bruker] scanner.
 
 ## Custom pulse sequences
-Getting started with pulse sequence design using `pypulseq` is simple:
+Getting started with pulse sequence design using `PyPulseq` is simple:
 1. `pip install pypulseq` in your virtual environment (>=Python 3.6).
 2. First, define system limits in `Opts` and then create a `Sequence` object with it:
     ```python
@@ -72,7 +72,7 @@ need to be executed simultaneously:
     ```
 
 ## Contributing and Community guidelines
-`pypulseq` adheres to a code of conduct adapted from the [Contributor Covenant] code of conduct.
+`PyPulseq` adheres to a code of conduct adapted from the [Contributor Covenant] code of conduct.
 Contributing guidelines can be found [here][contrib-guidelines].
 
 ---
