### 2019-09-18 #1725 @sravan953
@grlee77 I have address your comments in the most recent commit. Please take a look!

More specifically regarding #5 in your comments: our future efforts involve incorporating SAR computation and real-time pulse sequence design.
