### 2019-09-20 #1725 @mathieuboudreau
>Contribution and authorship: Has the submitting author (@imr-framework) made major contributions to the software? Does the full list of paper authors seem appropriate and complete?

@imr-framework could you clarify what contribution(s) John Thomas Vaughan Jr. made to this project; I checked out the contributors to the source code  [here](https://github.com/imr-framework/pypulseq/graphs/contributors), and he doesn't appear on the list (whereas @tonggehua has made some commits and doesn't appear on the paper).
