### 2019-09-20 #1725 @mathieuboudreau
@imr-framework does there exist some kind of validator tool to evaluate if a generated *.seq file is (1) properly formatted, and (2) doesn't exceed vendor-imposed hardware limits. I don't expect that this exists within your software, since *.seq files have been developed previously, I'm wondering if there's an external tool that already does this?
