### 2019-09-20 #1725 @sravan953
@mathieuboudreau
Regarding contributions-
John Thomas Vaughan Jr. has been principally involved in the design and review of this project, like for [Virtual Scanner](https://github.com/imr-framework/virtual-scanner/issues/22#issuecomment-520032559). We have been using PyPulseq internally, and the one commit from @tonggehua was regarding [documentation](https://github.com/imr-framework/pypulseq/commit/5cdf7a0802b764691cf1e5b7a6ae707edcb9a796). The authors list was finalised after internal deliberations.

Regarding the formatting-
We utilised a diffcheck tool to ensure that the *.seq files generated by PyPulseq match with Pulseq 1.2.0's files. We performed these checks for the demo scripts that have been bundled. The only difference between *.seq files generated by PyPulseq and Pulseq 1.2.0 is the comment on line 2 in every *.seq file which indicates which tool generated the file.

Regarding an external validator tool-
The `check_timing()` and `test_report()` methods can be leveraged to ensure that the designed pulse sequence is in accordance with the specified gradient raster time limits. Subsequently, a successfully generated *.seq file might or might not play on the scanner, depending on whether the hardware limits were violated or not.
