### 2019-09-20 imr-framework/pypulseq#16 @sravan953
@grlee77 Can this issue be closed?
