### 2019-09-23 imr-framework/pypulseq#16 @grlee77
It does not hurt to keep the issue open as a reminder to add this in the future. You could remove the "review comment" label and add a label like "task" to indicate that the issue is task-related and not a bug. Ultimately, it is up to you whether you find it useful to keep this open or not, though.
