### 2019-09-23 #1725 @grlee77
> @grlee77 I have addressed your comments in the most recent commit. Please take a look!

Thanks, I have checked off the items related above that were addressed.

Regarding functionality, I had not checked the box because I did not verify on an actual scanner, but given that the Pulseq files look correctly formatted and the author has performed an offline comparison vs. the same demos in the Matlab package, I would trust that they work equivalently on actual hardware.

Re Automated testing: This has not been implemented, but the author has indicated future plans to do so in imr-framework/pypulseq#16. It is currently possible for users to manually test the basic functionality of the software by running the supplied examples.
