### 2019-09-23 imr-framework/pypulseq#16 @sravan953
Completely on board about the benefits of CI - will incorporate it in the future, thank you for your inputs.
