### 2019-09-24 #1725 @spinicist
Hello,

I've gone through the review checklist.

The example in README.md does not appear to work. There is a typo - `slew_unit='mT/m/s'` should be `slew_unit='T/m/s'`, but after fixing that there is a division by zero error in `make_sinc_pulse`. The example EPI sequence works fine, as does the GRE sequence but it is very slow to plot.

Similarly to the other authors I did not go as far as testing the sequences on an actual scanner, as (for me) that would involve installation of TOPPE which would take considerable time and I consider beyond the scope of this review. On the same lines, I question the ultimate usefulness of tools such as PyPulseq and TOPPE because of their lack of interaction with vendor UI and safety limits (as mentioned by other reviewers). For instance all GE vendor sequences include extensive software SAR checks, which are not included in PyPulseq/TOPPE. I hence foresee conflicts with local safety and ethics policies if researchers want to use such sequences for scanning subjects.

I do not think the above is reason to reject this paper as PyPulseq itself is clearly useful for prototyping, and I have ticked the Functionality box because I think the functions that Py Pulseq claims to have, have been demonstrated. I only raise it because I think the realistic use-cases are fairly limited, and hence I think the claim "We also envisage PyPulseq to be utilized for repeatability and reproducibility studies ... (multi-site, multi-vendor)" is slightly strong.
