### 2019-09-24 97fdd37
Update invalid gradient/slew rate error messages
