### 2019-09-24 0f5dd67
Fix example code in README.md
