### 2019-09-24 #1725 @sravan953
@spinicist Thanks for your comments, have made changes to the README based on your inputs.

Regarding conflicts with local safety and ethics policies-
We will soon incorporate SAR safety features into PyPulseq. @sairamgeethanath has a MATLAB version available at - https://github.com/imr-framework/sar4seq, which will soon get integrated into PyPulseq. We advocate leveraging PyPulseq as an open-source tool for prototyping pulse sequences, and do not advise subject scanning ([1](https://onlinelibrary.wiley.com/doi/abs/10.1002/mrm.26235), [2](https://www.sciencedirect.com/science/article/pii/S0730725X1830033X)).

For repeatability and reproducibility-
[OMEGA](https://github.com/imr-framework/mrf) by @qianenlin focuses on the repeatability of MRF using NIST phantom ([Jiang et al., MRM 2014](https://onlinelibrary.wiley.com/doi/abs/10.1002/mrm.26509)). A MATLAB version exists as reported in [Qian et al., ISMRM 2019](https://index.mirasmart.com/ISMRM2019/PDFfiles/0817.html). We are now looking at developing a PyPulseq version.

The [AMRI](https://github.com/imr-framework/amri/tree/amri) project leverages PyPulseq. Video recordings demonstrating PyPulseq’s functionality from AMRI are available on [YouTube](https://www.youtube.com/watch?v=F5cQZWIHnQU&t=45s).
