### 2019-09-24 #1725 @sairamgeethanath
@sravan953 @spinicist The [AMRI phantom](https://www.youtube.com/watch?v=ZToRLwXQW1A) video illustrates the point better about using different acquisition parameters to generate different *.seq* files using pyPulseq and using them to scan phantoms, for example.
