### 2019-10-06 #1725 @sairamgeethanath
> @mathieuboudreau @grlee77 @spinicist We also have AMRI-Scanner as an Online Service (AMRI-SOS). AMRI-SOS lets you upload a sequence file to Google Drive and request a scan. On our end, we run an AMRI-SOS script that downloads the uploaded *.seq file and runs it on a Siemens Prisma 3T. This way we can prove functionality. Thanks @sairamgeethanath for the idea!

https://docs.google.com/forms/d/e/1FAIpQLSf3MIC6bcOpV20-yYrqmO_OhK4wNgsKZu3oBqRYe4CLbzIT8A/viewform?fbzx=3238199011716171808
