### 2019-10-07 #1725 @spinicist
I think my points have been sufficiently addressed. I have no further comments.

I'm not actually sure how to mark my review as complete?
