### 2019-10-08 #1725 @Kevin-Mattheus-Moerman
@imr-framework here are some additional issues/comments in relation to your paper. Please make these required changes:

- [ ] Check that you've added all possible DOI's to your references
- [ ] The hyper links `[TOPPE](toppe)`, `[Bruker](bruker)`,  `[PyPulseq-gpi](pypulseq-gpi-branch)`, and `[GPI](gpilab)` do not link to valid web content. Can you remove these links or fix them with proper URLs?
- [ ] You mention `We have leveraged this tool for several published research works`. Adding some key citations to those works here would be appropriate.
- [ ] The statement of need section defines the acronym MRI again (`Magnetic Resonance Imaging (MRI)`), since it is already defined in the preceding section please simply use it here instead.
- [ ] Since MR is not defined (I know it stands for Magnetic Resonance but still) I suggest you use MRI throughout instead. Or also define MR.
- [ ] In `Further, harmonizing acquisition across...` please replace `Further` with `Furthermore`.
- [ ] Remove the `-` from `open-source` throughout
- [ ] I suggest adding a comma after Python in `...platforms in Python based on the Pulseq standard...`
- [ ] Please rephrase `The file format was designed keeping in mind design goals such as: human-readable, easily parsable, vendor independent, compact and low-level (Layton et al., 2017).`, to something like: `The file format was designed with the following design criteria in mind: human-readable, easily parsable, vendor independent, compact and low-level (Layton et al., 2017).`
- [ ] Please consider rephrasing: `It enables researchers and users to program pulse sequences in Python and export as a .seq file.` to: `It enables researchers and users to program pulse sequences in Python, and to export them as a .seq file.`
- [ ] the `About PyPulseq`section contains several sentences starting with `First,..`, `Second,...`, and `Third...`. Please change thes to `Firstly,...`, `Secondly,...`, and `Thirdly,...`.
- [ ] In "`This allows PyPulseq to be integrated`...", It is unclear what `this` refers to? Please clarify. Perhaps saying `These advantages allow...` (if I understood this correctly) fixes the issue.
- [ ] Please rephrase and remove the use of etc in "`...and related downstream reconstruction,
etc.`".
- [ ] Please add the and in the following sentence "`PyPulseq is aimed at MRI researchers focusing on pulse sequence design, image reconstruc-tion, [and] MR physics.`"
- [ ] Once all changes are implemented I'll recommend that your paper is accepted. At this point I ask you to double check the text yourself, in particular check that author list, the affiliations and also the acknowledgement section is accurate.
