### 2019-10-08 985516b
Fix paper.bib and paper.md
- Add possible DOIs to paper.bib
- Grammar edits to paper.md
