+++ paper.md
@@ -1,5 +1,5 @@
 ---
-title: 'pypulseq'
+title: 'PyPulseq'
 tags:
   - Python
   - MRI
@@ -27,36 +27,37 @@ bibliography: paper.bib
 Magnetic Resonance Imaging (MRI) is a critical component of healthcare. MRI data is acquired by playing a series of 
 radio-frequency and magnetic field gradient pulses. Designing these pulse sequences requires knowledge of specific 
 programming environments depending on the vendor hardware (generations) and software (revisions) intended for 
-implementation. This impedes the pace of prototyping. Pulseq [@layton2017pulseq] introduced an open-source file 
-standard for pulse sequences that can be deployed on Siemens/GE via [TOPPE](toppe) [@nielsen2018toppe]/[Bruker](bruker) 
-platforms. In this work, we introduce `PyPulseq`, which enables pulse sequence programming in Python. Its advantages 
-are zero licensing fees and easy integration with deep learning methods developed in Python. `PyPulseq` is aimed at MR 
-researchers, faculty, students, and other allied field researchers such as those in neuroscience. We have leveraged 
-this tool for several published research works.
+implementation. This impedes the pace of prototyping. Pulseq [@layton2017pulseq] introduced an open source file 
+standard for pulse sequences that can be deployed on Siemens/GE via [TOPPE](https://toppemri.github.io) 
+[@nielsen2018toppe]/[Bruker](https://github.com/pulseq/bruker_interpreter) platforms. In this work, we introduce 
+`PyPulseq`, which enables pulse sequence programming in Python. Its advantages are zero licensing fees and easy 
+integration with deep learning methods developed in Python. `PyPulseq` is aimed at MRI researchers, faculty, students, 
+and other allied field researchers such as those in neuroscience. We have leveraged this tool for several published 
+research works [@poojar2019rapid; @gehua2019ismrm; @ravi2018amri; @ravi2019accessible-amri; @ravi2018imrframework; 
+@ravi2019selfadmin].
 
 # Statement of need
 
-Magnetic Resonance Imaging (MRI) is a non-invasive diagnostic imaging tool. It is a critical component of healthcare 
-and has a significant impact on diagnosis and treatment assessment. Structural, functional and metabolic MRI generate 
-valuable information that aid in the accurate diagnosis of a wide range of pathologies. A unique strength of MRI is the 
-ability to visualise diverse pathologies achieved by the flexibility in designing tailored pulse sequences. MR pulse 
-sequences are a collection of radio-frequency and gradient waveforms that are executed on the scanner hardware to 
-acquire raw data. 
+MRI is a non-invasive diagnostic imaging tool. It is a critical component of healthcare and has a significant impact on 
+diagnosis and treatment assessment. Structural, functional and metabolic MRI generate valuable information that aid in 
+the accurate diagnosis of a wide range of pathologies. A unique strength of MRI is the ability to visualise diverse 
+pathologies achieved by the flexibility in designing tailored pulse sequences. MRI pulse sequences are a collection of 
+radio-frequency and gradient waveforms that are executed on the scanner hardware to acquire raw data. 
 
 Research efforts on pulse sequence design are directed at achieving faster scan times, improving tissue contrast and 
 increasing Signal-to-Noise Ratio (SNR). However, designing pulse sequences requires knowledge of specific programming 
 environments depending on the vendor hardware (generations) and software (revisions) intended for implementation. 
-Typically, MRI researchers program and simulate the pulse sequences on computers and execute them on MR scanners. This 
+Typically, MRI researchers program and simulate the pulse sequences on computers and execute them on MRI scanners. This 
[…]
