+++ paper.bib
@@ -1,5 +1,6 @@
 @article{layton2017pulseq,
   title={Pulseq: a rapid and hardware-independent pulse sequence prototyping framework},
+  doi={10.1002/mrm.26235},
   author={Layton, Kelvin J and Kroboth, Stefan and Jia, Feng and Littin, Sebastian and Yu, Huijun and Leupold, Jochen and Nielsen, Jon-Fredrik and St{\"o}cker, Tony and Zaitsev, Maxim},
   journal={Magnetic resonance in medicine},
   volume={77},
@@ -27,6 +28,7 @@
 
 @article{ravi2018pulseq-gpi,
   title={Pulseq-Graphical Programming Interface: Open source visual environment for prototyping pulse sequences and integrated magnetic resonance imaging algorithm development},
+  doi={10.1016/j.mri.2018.03.008},
   author={Ravi, Keerthi Sravan and Potdar, Sneha and Poojar, Pavan and Reddy, Ashok Kumar and Kroboth, Stefan and Nielsen, Jon-Fredrik and Zaitsev, Maxim and Venkatesan, Ramesh and Geethanath, Sairam},
   journal={Magnetic resonance imaging},
   volume={52},
@@ -45,6 +47,7 @@
 
 @article{jochimsen2004odin,
   title={ODIN—object-oriented development interface for NMR},
+  doi={10.1016/j.jmr.2004.05.021},
   author={Jochimsen, Thies H and Von Mengershausen, Michael},
   journal={Journal of Magnetic Resonance},
   volume={170},
@@ -56,6 +59,7 @@
 
 @article{magland2016pulse,
   title={Pulse sequence programming in a dynamic visual environment: SequenceTree},
+  doi={10.1002/mrm.25640},
   author={Magland, Jeremy F and Li, Cheng and Langham, Michael C and Wehrli, Felix W},
   journal={Magnetic resonance in medicine},
   volume={75},
@@ -67,6 +71,7 @@
 
 @article{nielsen2018toppe,
   title={TOPPE: A framework for rapid prototyping of MR pulse sequences},
+  doi={10.1002/mrm.26990},
   author={Nielsen, Jon-Fredrik and Noll, Douglas C},
   journal={Magnetic resonance in medicine},
   volume={79},
@@ -74,6 +79,14 @@
   pages={3128--3134},
   year={2018},
   publisher={Wiley Online Library}
+}
+
+@article{poojar2019rapid,
+  title={Rapid prOtotyping of 2D non-CartesIan K-space trajEcTories (ROCKET) using Pulseq and GPI},
+  doi={10.1615/CritRevBiomedEng.2019029380},
+  author={Poojar, Pavan and Geethanath, Sairam and Reddy, Ashok Kumar and Venkatesan, Ramesh},
+  journal={Critical Reviews™ in Biomedical Engineering},
+  publisher={Begel House Inc.}
 }
 
 @inproceedings{ravi2019accessible-amri,
