### 2019-10-09 #1725 @Kevin-Mattheus-Moerman
@sravan953 thanks. One point is remaining, in the sentence containing"`...and export them a .seq file.`", you forgot to add either `to` or `as` i.e. use "`...and export them [to] a .seq file.`" or "`...and export them [as] a .seq file.`"
