### 2019-10-09 #1725 @Kevin-Mattheus-Moerman
@sravan953, once you've made the above change, can you run `@whedon generate pdf` here? Also can you do the following:

- [ ] Please archive the reviewed software on Zenodo and report the DOI here in this review thread. Make sure the Zenodo deposit has the correct metadata (same title and author list as the paper).
- [ ] Can you confirm that the version of the reviewed (and archived) software is `1.2.2, or has the version tag moved on? Please provide the latest version tag for the reviewed software.
