### 2019-10-09 3d72091
Fix grammar in paper.md
