### 2019-10-10 #1725 @sravan953
@Kevin-Mattheus-Moerman -
1. DOI - 10.5281/zenodo.3479527
2. Latest release - 1.2.2r1 (published on Zenodo)
