### 2019-10-11 #1725 @Kevin-Mattheus-Moerman
@sravan953 one last point from me. I would recommend that you update your paper title (and also the matched Zenodo archive title) to be a bit more descriptive, e.g.:
**PyPulseq: A Python Package for MRI Pulse Sequence Design**
Something like that would help make your paper more discoverable.
