### 2019-10-11 #1725 @arfon
> @arfon is this something that we can change just in the PDF or do we need to propagate the title change anywhere else?

Just the `paper.md` should do it.
