### 2019-10-11 963bea1
Edit title in paper.md and README.md
