### 2019-10-11 7972987
Remove WIP sequences
