+++ pypulseq/seq_examples/wip/test.py
@@ -1,5 +1,0 @@
-from pypulseq.Sequence.sequence import Sequence
-
-seq = Sequence()
-seq.read('/Users/sravan953/Downloads/25-Sep-2018S1_T1.seq')
-seq.duration()

[…]
