### 2019-10-11 #1725 @sravan953
@Kevin-Mattheus-Moerman @arfon Updated the title on the `README.md`, Zenodo archive and on `paper.md`.
