### 2019-10-11 #1725 @labarba
I fixed in-text citation syntax and a terminology issue in https://github.com/imr-framework/pypulseq/pull/23
