### 2019-10-12 #1725 @labarba
Huge thanks to our editor: @Kevin-Mattheus-Moerman, and the reviewers: @grlee77, @mathieuboudreau, @spinicist — your contributions to JOSS are greatly appreciated 🙏
