### 2019-10-14 #1725 @sairamgeethanath
> Congratulations, @imr-framework, your JOSS paper is published! 🚀
>
> Huge thanks to our editor: @Kevin-Mattheus-Moerman, and the reviewers: @grlee77, @mathieuboudreau, @spinicist — your contributions to JOSS are greatly appreciated 🙏

Many thanks to all the reviewers @mathieuboudreau @spinicist @grlee77, the editor @Kevin-Mattheus-Moerman and the associate editor in chief @labarba for the time and inputs to make our work better.
